+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-10-19
+
+### Added
+
+- new command `Export-DataJourney`.
+- new command `Import-DataJourney`.
+- new command `New-DataFlow`.
+- new command `New-DataJourney`.
+- new command `New-DataLayer`.
+- new command `New-DataModel`.
+- new command `Select-DataJourney`.
+- new command `Set-DataJourney`.
+
 ## [0.1.0] - 2023-09-14
 
 ### Added